splits firewall functionality into another file
